Add IPCParsedCommand and refactor command parsing
First, use unsigned long and long for the Arg union, so a memory address will
fit as an unsigned long.

Create an IPCParsedCommand struct that contains an array of Arg, the number of
arguments, the argument types, and the command name which can all be populated
in ipc_parse_run_command after extracting this information from the client's
message. This enables the separation of argument checking and argument
extraction and also cleans up the parsing funcion.

ipc_validate_parsed_command compares IPCCommand to IPCParsedCommand and checks
if the number of arguments and argument types match. For two specific cases, the
argument types don't have to match:
- When the command expects a signed int, but an unsigned int is given
- When the command expects a void pointer, but an unsigned int is given
In both these cases, the arguments are casted to the correct argument and
reassigned to the union.

ipc_free_parsed_command_members frees the members of the struct taking into
consideration if the argument type is ARG_TYPE_STR, in which case that
particular element of args has to also be freed.

ipc_parse_run_command now only focuses on parsing the commands.

ipc_run_command uses ipc_parse_run_command to parse the commands,
ipc_get_ipc_command to check if the command exists, and
ipc_validate_parsed_command to validate the command.

IPCCommand.command_name was also renamed to IPCCommand.name for simplicity.

+++ types.h
@@ -4,8 +4,8 @@
 #include <X11/Xlib.h>
 
 typedef union {
-  int i;
-  unsigned int ui;
+  long i;
+  unsigned long ui;
   float f;
   const void *v;
 } Arg;
